requirements.txt
streamlit
python-dotenv
phidata
requests
rich
openai